Strip excluded source map references
